+++ api/bin/app.php
@@ -1,0 +1,46 @@
+#!/usr/bin/env php
+<?php
+declare(strict_types=1);
+
+use Doctrine\Migrations\Configuration\Configuration;
+use Doctrine\Migrations\Tools\Console\ConsoleRunner as MigrationConsoleRunner;
+use Doctrine\Migrations\Tools\Console\Helper\ConfigurationHelper;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Tools\Console\ConsoleRunner;
+use Doctrine\ORM\Tools\Console\Helper\EntityManagerHelper;
+use Psr\Container\ContainerInterface;
+use Symfony\Component\Console\Application;
+use Symfony\Component\Dotenv\Dotenv;
+
+chdir(dirname(__DIR__));
+require 'vendor/autoload.php';
+
+if (file_exists('.env')) {
+    (new Dotenv())->load('.env');
+}
+
+/** @var ContainerInterface $container */
+$container = require 'config/container.php';
+
+$cli = new Application('Application console');
+
+$entityManager = $container->get(EntityManagerInterface::class);
+$connection = $entityManager->getConnection();
+
+$configuration = new Configuration($connection);
+$configuration->setMigrationsDirectory('src/Data/Migration');
+$configuration->setMigrationsNamespace('Api\Data\Migration');
+
+$cli->getHelperSet()->set(new EntityManagerHelper($entityManager), 'em');
+$cli->getHelperSet()->set(new ConfigurationHelper($connection, $configuration), 'configuration');
+
+ConsoleRunner::addCommands($cli);
+MigrationConsoleRunner::addCommands($cli);
+
+$commands = $container->get('config')['console']['commands'];
+
+foreach ($commands as $command) {
+    $cli->add($container->get($command));
+}
+
+$cli->run();