+++ api/src/Model/User/UseCase/SignUp/Request/Handler.php
@@ -1,0 +1,57 @@
+<?php
+declare(strict_types=1);
+
+namespace Api\Model\User\UseCase\SignUp\Request;
+
+use Api\Model\User\Entity\User\Email;
+use Api\Model\User\Entity\User\User;
+use Api\Model\User\Entity\User\UserId;
+use Api\Model\User\Entity\User\UserRepository;
+use Api\Model\User\Flusher;
+use Api\Model\User\Service\ConfirmTokenizer;
+use Api\Model\User\Service\PasswordHasher;
+
+class Handler
+{
+    /** @var UserRepository */
+    private $users;
+    /** @var PasswordHasher */
+    private $hasher;
+    /** @var ConfirmTokenizer */
+    private $tokenizer;
+    /** @var Flusher */
+    private $flusher;
+
+    public function __construct(
+        UserRepository $users,
+        PasswordHasher $hasher,
+        ConfirmTokenizer $tokenizer,
+        Flusher $flusher
+    ){
+        $this->users = $users;
+        $this->hasher = $hasher;
+        $this->tokenizer = $tokenizer;
+        $this->flusher = $flusher;
+    }
+
+    public function handle(Command $command): void
+    {
+        $email = new Email($command->email);
+
+        if ($this->users->hasByEmail($email)) {
+            throw new \DomainException('User with this email already exists.');
+        }
+
+        $user = new User(
+            UserId::next(),
+            new \DateTimeImmutable(),
+            $email,
+            $this->hasher->hash($command->password),
+            $token = $this->tokenizer->generate()
+        );
+
+        $this->users->add($user);
+
+        $this->flusher->flush();
+    }
+}