+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
-All notable changes to this **shop** project will be documented in this file. 
+All notable changes to this **shop** project will be documented in this file.
+
+## [0.0.2] - 2020-03-28
+### Added
+- User domain
+- api endpoints for registration and confirm registration of user
+- libraries ramsey/uuid, doctrine/orm, zendframework/zend-config-aggregator, symfony/console, doctrine/migrations, doctrine/data-fixtures, symfony/validator
+- container services postgres, php-cli
+- cli api project
+- cli commands into composer launcher
+- migrations
+- fixtures
+- validator
+- middlewares
+### Updated
+- config of api project
 
 ## [0.0.1] - 2020-03-14
 ### Added